1、通过将资源加载到内存中，加速游戏启动速度； 2、资源加载过程放到线程中，主界面显示加载进度； 3、Android版支持切换到后台时将游戏进度保存到临时文件中，若系统杀死游戏后再启，可选择继续游戏； 4、调整资源文件存储结构 5、设置Windows版的应用图标，与Android版保持一致； 6、修正暂停后可通过任一按钮进入游戏的bug；

+++ core/Animation.cpp
@@ -281,7 +281,7 @@ shared_ptr<Animation> Animation::build(void){
     newAnimation->setCalculateHandlerX(m_calculateHandlerX);
     newAnimation->setCalculateHandlerY(m_calculateHandlerY);
     newAnimation->setCalculateHandlerAlpha(m_calculateHandlerAlpha);
-    newAnimation->setStartFrame(m_currentFrame);
+    newAnimation->setStartFrame((int)m_currentFrame);
     newAnimation->setLoop(m_isLoop);
     newAnimation->setInitialPos(m_targetX, m_targetY);
     newAnimation->setInitialAlpha(m_targetAlpha);
